refactor: 테스트코드마다 Displayname 작성
