+++ src/test/java/carracing/model/CarsTest.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.model;
 
 import static carracing.view.messages.ExceptionMessage.*;
 import static org.assertj.core.api.Assertions.*;
@@ -8,30 +8,11 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
-public class CarRacingTest {
+public class CarsTest {
 
 	@Test
-	@DisplayName("자동차 이름의 길이가 5를 초과하는 경우 에러 발생")
-	void validate_자동차이름_길이() {
-		assertThatThrownBy(() -> {
-			new Car("pobbiii");
-		}).isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining(CAR_NAME_LENGTH_EXCEPTION.getMessage());
-	}
-
-	@Test
-	@DisplayName("자동차 이름의 길이가 1미만인 경우 에러 발생")
-	void validate_자동차이름_길이기준_미달() {
-			assertThatThrownBy(() -> {
-				new Car("");
-			}).isInstanceOf(IllegalArgumentException.class)
-				.hasMessageContaining(CAR_NAME_LENGTH_EXCEPTION.getMessage());
-	}
-
-	@Test
+	@DisplayName("자동차의 이름이 중복될 경우 에러 발생")
 	void validate_자동차이름_중복(){
 		//given
 		//when
@@ -47,32 +28,8 @@ public class CarRacingTest {
 			.hasMessageContaining(CAR_NAME_DUPLICATE_EXCEPTION.getMessage());
 	}
 
-	@ParameterizedTest
-	@ValueSource(ints = {4, 5, 6, 7, 8, 9})
-	void 랜덤값_4이상이면_전진(int randomNumber) {
-		//given
-		Car car = new Car("pobi");
-		//when
-		car.move(randomNumber);
-		int position = car.getPosition();
-		//then
-		assertThat(position).isEqualTo(1);
-	}
-
-	@ParameterizedTest
-	@ValueSource(ints = {0, 1, 2, 3})
-	void 랜덤값_4미만이면_멈춤(int randomNumber) {
-		//given
-		Car car = new Car("pobi");
-		int originalPosition = car.getPosition();
-		//when
-		car.move(randomNumber);
-		int changedPosition = car.getPosition();
-		//then
-		assertThat(changedPosition).isEqualTo(originalPosition);
-	}
-
 	@Test
+	@DisplayName("position이 가장 높은 자동차가 우승")
 	void 최종우승자_결정_한명() {
 		//given
 		Car pobiCar = new Car("pobi");
@@ -96,6 +53,7 @@ public class CarRacingTest {
 	}
 
 	@Test
+	@DisplayName("position이 가장 높은 자동차가 두 대일 때 공동 우승자")
 	public void 최종우승자_결정_여러명() throws Exception{
 		//given
 		Car pobiCar = new Car("pobi");
